braincoins/braincoins.old/requirements.txt
annotated-types==0.6.0
appdirs==1.4.4
asttokens==2.4.1
attrs==23.1.0
audioread==3.0.1
beartype==0.16.4
bleach==6.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
decorator==5.1.1
deepspeed==0.8.3
einops==0.4.1
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.0
ffmpeg==1.4
filelock==3.13.1
fsspec==2023.12.2
hjson==3.1.0
huggingface-hub==0.20.1
idna==3.6
inflect==7.0.0
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter_core==5.5.1
lazy_loader==0.3
librosa==0.9.1
llvmlite==0.41.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mistune==3.0.2
mpmath==1.3.0
msgpack==1.0.7
mysql-connector-python==8.2.0
nbconvert==5.3.1
nbformat==5.9.2
networkx==3.2.1
ninja==1.11.1.1
numba==0.58.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
packaging==23.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.9.0
pika==1.3.2
platformdirs==4.1.0
pooch==1.8.0
progressbar==2.5
prompt-toolkit==3.0.43
protobuf==4.21.12
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pycparser==2.21
pydantic==1.9.1
pydantic_core==2.14.5
Pygments==2.17.2
PyYAML==6.0.1
referencing==0.32.0
regex==2023.10.3
requests==2.31.0
resampy==0.4.2
rotary-embedding-torch==0.3.6
rpds-py==0.15.2
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
sounddevice==0.4.6
soundfile==0.12.1
soxr==0.3.7
stack-data==0.6.3
sympy==1.12
testpath==0.6.0
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.1.2
torchaudio==2.1.2
tornado==4.2
tortoise-tts==3.0.0
tqdm==4.66.1
traitlets==5.14.0
transformers==4.31.0
triton==2.1.0
typing_extensions==4.9.0
Unidecode==1.3.7
urllib3==2.1.0
wcwidth==0.2.12
webencodings==0.5.1